* LunarMonthUnit.getDurationInDays 반환꼴 double로 변경: 소수점 밑 자르면 안 됐음.(정밀도 너무 떨어짐) * KLunarDate.resolveClosestDayOfMonth 계산식 오류

+++ src/main/java/xyz/arinmandri/koreanlunarcalendar/LunarMonthUnit.java
@@ -35,8 +35,8 @@ public enum LunarMonthUnit implements TemporalUnit
 		return duration;
 	}
 
-	public int getDurationInDays () {
-		return ( (int) duration.getSeconds() ) / ( 24 * 60 * 60 );
+	public double getDurationInDays () {
+		return ( (double) duration.getSeconds() ) / ( 24 * 60 * 60 );
 	}
 
 	@Override
